pyproject.toml
[tool.poetry]
name = "temperature"
version = "0.1.0"
description = ""
authors = ["Larisa Shigareva <thelarisa81@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
Flask = "^1.1.2"
celery = "^5.0.5"
python-dotenv = "^0.17.1"
requests = "^2.25.1"
redis = "^3.5.3"
icecream = "^2.1.0"
python-dateutil = "^2.8.2"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
